# backend/documentation/swagger.yml
swagger: "2.0"
info:
  description: "Die API-Dokumentation für [current](https://current.land)."
  version: "1.0.0"
  title: "current API"
  contact:
    email: "hi@malts.me"
host: "api.current.land"
basePath: "/"
...
schemes:
- "https"
paths:
  /posts:
    post:
      tags:
      - "posts"
      summary: "Create a new post"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Post object that needs to be added to the store"
        required: true
        schema:
          type: "object"
          properties:
            title:
              type: "string"
              example: "exampleuser"
            markdownBody:
              type: "string"
              example: "*hello* world!"
            thubmnail:
              type: "string"
              example: "507f191e810c19729de860ea"
      responses:
        "400":
          description: "Invalid input"
    get:
      summary: "Get posts"
      tags:
      - "posts"
      parameters:
      - name: "skip"
        in: "query"
        description: "amount of posts to skip"
        type: "integer"
      - name: "limit"
        in: "query"
        description: "limit the amount of posts to a given value"
        type: "integer"
      - name: "user"
        in: "query"
        description: "only get posts of specified user"
        type: "string"
      responses:
        "200":
          description: "success"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "error"
          schema:
            type: "object"
            properties:
              error:
                type: "object"
                properties:
                  message:
                    type: string
                    example: 'invalid username'
  /posts/{postId}:
    get:
      summary: "Get a single post"
      tags:
      - "posts"
      parameters:
      - name: "postId"
        in: "path"
        description: "Id of the post"
        required: true
        type: "integer"
      responses:
        "200":
          description: "success"
          schema:
            $ref: "#/definitions/Post"
        "404":
          description: "error"
          schema:
            type: "object"
            properties:
              error:
                type: "object"
                properties:
                  message:
                    type: string
                    example: 'Post not found'
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Post:
    type: "object"
    properties:
      title:
        type: "string"
      markdownBody:
        type: "string"
      thumbnail:
        type: "string"
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"
      userStatus:
        type: "integer"
        format: "int32"
        description: "User Status"
    xml:
      name: "User"
...